Started to create the structure for a summary/tldr component
