+++ apps/www/src/components/summary/index.ts
@@ -1,0 +1,2 @@
+export { Summary } from './summary'
+export type { SummaryProps } from './types'
