+++ apps/www/src/components/summary/types.ts
@@ -1,0 +1,6 @@
+type SummaryProps = {
+  title: string,
+  description?: any,
+}
+
+export type { SummaryProps }
